backend/requirements.txt
# FastAPI and Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
pydantic==2.6.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
asyncpg==0.29.0
pgvector==0.2.4

# OpenAI and LangChain
openai==1.12.0
langchain==0.1.6
langchain-community==0.0.20
langchain-openai==0.0.5
tiktoken==0.6.0

# Vector Store and Embeddings
sentence-transformers==2.3.1
faiss-cpu==1.7.4
chromadb==0.4.22

# Voice Processing
google-cloud-speech==2.24.0
google-cloud-texttospeech==2.15.0
azure-cognitiveservices-speech==1.34.1

# Arabic Text Processing
arabic-reshaper==3.0.0
python-bidi==0.4.2
pyarabic==0.6.15

# PDF and Document Processing
pypdf==4.0.1
pymupdf==1.23.21
python-docx==1.1.0

# HTTP and Async
httpx==0.26.0
aiohttp==3.9.3
requests==2.31.0

# Utilities
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
celery==5.3.6

# Logging and Monitoring
loguru==0.7.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
